requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Environment and Configuration
python-dotenv==1.0.0

# LLM Integration
openai==1.12.0

# Embeddings and Vector Store
sentence-transformers==2.3.1
faiss-cpu

# Document Parsing
PyPDF2==3.0.1
markdown==3.5.2

# Utilities
python-multipart==0.0.9

# Caching (optional)
redis==5.0.1
django-redis==5.4.0

# Development
pytest==7.4.4
pytest-django==4.7.0
